don't write expected number of data points to metadata
When we write the number of expected data points into the HDF5 file,
this has the effect that datasets with the corresponding amount of rows
are created (all values being 0). This is not correct. Instead, we now
update the number of data points every time we add a data point and
resize the dataset accordingly.