PreProcessing: fix parameter_dict usage
The parameter_dict of the pre-processing worker was updated in the wrong
place (i.e. after the json was generated). But instead of updating it, I
should just pass the data point values to the generate_json_sequence
function, overwriting the parameter_dict values. Otherwise, device
parameters might end up in the parameter_dict, which don't belong there.