adds scannable_params to the device information dict
When a client requests the device information, it will get an additional
scannable_params list containing the full_access_path of all parameters
that can be scanned on the device. If the device structure changes, the
devices_callback will emit an update. This is added as a callback to the
observer within the IconServer.post_startup hook.