JobRepository: fixes queries, expunging objects
When log level is INFO (or higher), the job objects where not loaded and
the accessing attributes of the job instance would raise a
DetachedInstanceError in the scheduler. Expunging the instances helps
with that. However, with joined loads, I need to separate the update and
the query statements for the joined load to work.