HardwareProcessing: adds created attribute to HardwareProcessingTask
When multiple hardware tasks from the same pre-processing worker are
queued, the order in which they are taken from the queue is
non-deterministic (it seems).
Adding a "created" attribute to the hardware task and using it to define
the priority in which they are taken from the queue fixes this.
Also, updates type of processed_data_points