frontend: updates JobViewerPage
- uses extractStorageKey to get the localStorage key for the window
bounds. If the window name is set and starts with "jobView:", the window
was opened by either double-clicking a job in the data page, or
submitting a new job through the experiments page. If so, the window
should take the bounds stored in the localStorage
- on moving/resizing the JobViewer page, it saves its bounds to the
local storage (under its storageKey)
- adds a document title if the storageKey is defined