fix: notify clients about metadata changes
Before this change, the window that was opened when starting an
experiment would not show any plots. The reason is that the HDF file was
not created when the data was requested, meaning the metadata was also
not available yet.
This change notifies clients about metadata changes, so the window that
was opened registers to the metadata event and gets the update once its
available.